solar_panel_app/requirements.txt
streamlit
torch
torchvision
pillow
matplotlib
numpy
pycocotools
